Initial feature-complete version!
Major rewrite of the JobDatabase and JobQueue classes, to allow for job dependencies to be specified and resolved in the proper order (such as an EncodeJob waiting for a RipJob to finish before starting).

Full support for series with variable numbers of episodes has been completed.

Tons of refactoring and streamlining throughout the code to decouple classes and make better use of dependency injection and polymorphism to allow for future extensibility.